requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.11
djangorestframework==3.14.0
et-xmlfile==1.1.0
jdcal==1.4.1
openpyxl==2.4.11
pytz==2024.1
sqlparse==0.4.4
typing-extensions==4.10.0
XlsxWriter==3.2.0
